Add automatic management of snapshots (deletion and load)
This commit adds two new feature to snapshot system of Chainer. The
first feature is automatic deletion of old snapshot files during the
training. This is done by adding a post-save hook interface to
snapshot writers and if the number of snapshots to retain is set as
positive integers, the snapshot writers automatically lists all files
in Trainer output directory and matches the filename format.  One
limitation is that the automatic matching is only for integers but
does not work for arbitrary formatting of objects or string. For
example, if the filename format is defined as
`snapshot_iter_{.updater.iteration}` , the automatic matching system
picks files named as `snapshot_iter_10`, `snapshot_iter_12000` and so
on.

Another feature is automatic loading of snapshot target at extension
initialization. If `autoload` option is set as True it scans the
Trainer output directory and finds matches the filename format. For
example, if files named as `snapshot_iter_10` and
`snapshot_iter_12000` are found at startup, it chooses the latter
as the latest snapshot and loads the data to target.

This commit also adds two newly exposed function to snapshot extension,
`find_latest_snapshot` and `find_stale_snapshots`.